# api-docs.json
{"openapi":"3.0.1","info":{"title":"Products Management API","description":"This API exposes endpoints to manage products.","termsOfService":"https://www.netmind.es/terms","contact":{"name":"Contact name","url":"https://www.netmind.es","email":"contact@c.com"},"license":{"name":"MIT License","url":"https://choosealicense.com/licenses/mit/"},"version":"1.0"},"servers":[{"url":"http://localhost:9090","description":"Server URL in Development environment"},{"url":"https://localhost:8443","description":"Server URL in Production environment"}],"security":[{"Bearer Authentication":[]}],"tags":[{"name":"API de productos","description":"Esta es una api para crear y consumir productos"}],"paths":{"/products/{pid}":{"get":{"tags":["API de productos"],"summary":"Obtener un producto","description":"Endpoint para obtener un producto concreto","operationId":"getOne","parameters":[{"name":"pid","in":"path","description":"El identificador del producto","required":true,"schema":{"minimum":1,"type":"integer","format":"int64"},"example":234}],"responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"Todo bien","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Producto"}}}},"404":{"description":"El producto no existe","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Producto"}}}}}},"put":{"tags":["API de productos"],"operationId":"updateProduct","parameters":[{"name":"pid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Producto"}}},"required":true},"responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["API de productos"],"operationId":"deleteProduct","parameters":[{"name":"pid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/users":{"get":{"tags":["user-service-controller"],"operationId":"getAllUsers","responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-service-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/products":{"get":{"tags":["API de productos"],"operationId":"getAll","parameters":[{"name":"text","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Producto"}}}}}}},"post":{"tags":["API de productos"],"summary":"Crear un producto","description":"Endpoint para crear un producto concreto","operationId":"create","requestBody":{"description":"El producto a crear sin Id.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Producto"}},"application/atom+xml":{"schema":{"$ref":"#/components/schemas/Producto"}}},"required":true},"responses":{"412":{"description":"El producto enviado no es correcto","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"202":{"description":"Todo bien","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Producto"}}}}}}},"/auth/login":{"post":{"tags":["auth-service"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"412":{"description":"Precondition Failed","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/StatusMessage"},{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"StatusMessage":{"type":"object","properties":{"status_code":{"type":"integer","format":"int32"},"message":{"type":"string"}}},"Producto":{"required":["name","serial"],"type":"object","properties":{"Id":{"type":"integer","description":"El identificador único del producto. Debe ser un número mayor que 0. No es requerido en POST y PUT","format":"int64","example":234},"name":{"maxLength":20,"minLength":3,"type":"string"},"serial":{"type":"string"}},"description":"Representa un producto a intercambiar"},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","GESTOR","CLIENTE"]}}},"AuthRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"Bearer_Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}